stars v1.0
"templates/stars" contains a set of panels for the use of STARS
in VnmrJ 1.1. The original interface for VNMR (the contents of
"templates/stars61b.tar.Z") is also present with a few bug fixes. The
VnmrJ panels are contained in a fourth tab "ANALYZE". They are named
"STARSsetup" and "STARSdata". Also included is a macro "rtcmx" for
loading Spinsight data along with the conversion program "convertcmx".